Reference files without need to symlink or copy them.
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-libs-2.1.1@16 75748be7-0387-4326-94c7-f41ce0883d2c

+++ libsidutils/src/SidFilter.cpp
@@ -100,7 +100,7 @@ void SidFilter::read (ini_fd_t ini, const char *heading)
                 return;
             }
         }
-        *sidparams[i].address = tmp;
+        *sidparams[i].address = (float) tmp;
     }
 }
